ingestion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aio-pika==9.4.3
pydantic-settings==2.5.2
pydantic==2.9.2
python-jose[cryptography]==3.3.0
